Rename v2 to v2alpha1
