+++ pkg/chains/formats/all/all.go
@@ -17,6 +17,6 @@ package all
 import (
 	_ "github.com/tektoncd/chains/pkg/chains/formats/simple"
 	_ "github.com/tektoncd/chains/pkg/chains/formats/slsa/v1"
-	_ "github.com/tektoncd/chains/pkg/chains/formats/slsa/v2"
+	_ "github.com/tektoncd/chains/pkg/chains/formats/slsa/v2alpha1"
 	_ "github.com/tektoncd/chains/pkg/chains/formats/slsa/v2alpha2"
 )
