+++ pkg/chains/formats/format.go
@@ -32,7 +32,7 @@ const (
 	PayloadTypeSimpleSigning config.PayloadType = "simplesigning"
 	PayloadTypeInTotoIte6    config.PayloadType = "in-toto"
 	PayloadTypeSlsav1        config.PayloadType = "slsa/v1"
-	PayloadTypeSlsav2        config.PayloadType = "slsa/v2alpha1"
+	PayloadTypeSlsav2alpha1  config.PayloadType = "slsa/v2alpha1"
 	PayloadTypeSlsav2alpha2  config.PayloadType = "slsa/v2alpha2"
 )
 
@@ -40,7 +40,7 @@ var (
 	IntotoAttestationSet = map[config.PayloadType]struct{}{
 		PayloadTypeInTotoIte6:   {},
 		PayloadTypeSlsav1:       {},
-		PayloadTypeSlsav2:       {},
+		PayloadTypeSlsav2alpha1: {},
 		PayloadTypeSlsav2alpha2: {},
 	}
 	payloaderMap = map[config.PayloadType]PayloaderInit{}
